+++ features/chat/messageOnDrop.js
@@ -1,3 +1,4 @@
+import settings from '../../settings';
 import { getDropMessage } from '../../utils/common';
 import { isInSkyblock } from '../../utils/playerState';
 
@@ -9,8 +10,10 @@ export function sendMessageOnDrop(options) {
 			return;
 		}
 		
-		const message = getDropMessage(options.itemName);
-		ChatLib.command(chatCommand + ' ' + message);	
+		const message = settings.includeMagicFindIntoDropMessage
+			? getDropMessage(options.itemName, options.magicFind)
+			: getDropMessage(options.itemName);
+		ChatLib.command(chatCommand + ' ' + message);
 	} catch (e) {
 		console.error(e);
 		console.log(`[FeeshNotifier] Failed to send the message and play alert on drop.`);